added "no eval" if /etc/pluto.conf does not exist so it doesn't return an error

+++ BootScripts/Config_Ops.sh
@@ -4,6 +4,7 @@ NoSpace="s/ //g"
 
 ConfEval()
 {
+	[ -e /etc/pluto.conf ] || exit 0
 	sed "$NoSpace" /etc/pluto.conf >/tmp/pluto.conf
 	while read line; do
 		eval "export $line"
